digit-recognizer/Code/Web-app/requirements.txt
keras==2.9.0
matplotlib==3.5.2
numpy==1.22.3
pandas==1.4.2
Pillow==9.1.1
streamlit==1.9.0
tensorflow==2.9.0
